feat(network): the whole 5G surface exists only on a box that has the modem
The Photonicat 2 ships in variants, and until now the 5G half was offered on
every box. That is a real foot-gun, not just clutter: on a box with no modem,
picking `5g` suppresses the ethernet default route with nothing to replace it.
Recoverable — NFR-NET-01 keeps the cockpit on the LAN address and the fleet ULA
— but it should never have been reachable. New FR-NET-16.

What was already hardware-gated: `status.wwan` (null without a modem). What was
not: the bearer configuration, and — the dangerous one — the `wan_5g` and `5g`
mode options.

* Presence is probed from **sysfs** (`/sys/class/usbmisc/cdc-wdm*`,
  `/sys/class/net/ww*`), NOT from ModemManager. `mmcli` answering is a statement
  about a daemon's current view; using it would make the whole 5G surface appear
  and disappear across an MM restart. `CP_WWAN_PRESENT=0|1` overrides it for a
  variant the probe reads wrong.
* Server-side: `POST …/mode` with `wan_5g`/`5g` and `POST …/wwan` are `400` on a
  box with no modem; `config.wwan` is null; the applier creates no `cp-wwan`
  profile (an orphan connection bound to an absent device is just something for
  a human to puzzle over later).
* `status.modem_present` is the new hardware fact, readable by any
  `can_manage_it` caller — choosing the uplink mode is the client admin's job
  even though the bearer's configuration is the vendor's.
* The cockpit renders only the ethernet option and says why: "This box has no 5G
  module, so ethernet is its only uplink."
* Ansible probes the same way and skips the modem toolbox on a box without one —
  four packages and a ModemManager daemon that were being installed to probe
  ports that do not exist.

The hardware fact is threaded into the handlers as a parameter rather than
probed inside them, so the no-modem variant is testable without mutating the
environment (this workspace forbids `unsafe`, which `std::env::set_var` needs).

Structure: both panes crossed the 500-line ceiling and `config/` was at its
8-entry cap, so the IT category gets its own `config/it/` directory —
`ItPane`, `ItNetworkPane` and the new vendor-only `ItWwanForm`. `config/` drops
to 7. `Toggle` moves next to the other field primitives in `ItPane` so the pane
and the form can share it without a module cycle.

Verified on dh-7681f2a227e0f10d, the same box in both shapes:

  with modem   : modem_present=true,  mode=wan_5g → 200, all 3 options rendered,
                 bearer section rendered for a superadmin
  CP_WWAN_PRESENT=0 : modem_present=false, mode=5g/wan_5g → 400
                 "this box has no 5G modem", config.wwan null, only the ethernet
                 option rendered plus the explanation, no bearer section
