+++ web/src/components/shell/config/ConfigPanes.tsx
@@ -5,8 +5,8 @@ import { Check, Lock } from "lucide-react"
 import { UsagePage } from "@/components/agents/UsagePage"
 import { UpdatePane } from "./UpdatePane"
 import { SecretsPane } from "./SecretsPane"
-import { ItNetworkPane } from "./ItNetworkPane"
-import { ItPane } from "./ItPane"
+import { ItNetworkPane } from "./it/ItNetworkPane"
+import { ItPane } from "./it/ItPane"
 import { useProviders } from "@/lib/support/models"
 import { fetchSettings, updateSettings, fetchEnvKeys } from "@/lib/api"
 import { useAuth } from "@/lib/providers/auth"
